requirements.txt
pandas
numpy
matplotlib
opencv-python
torch
albumentations
segmentation-models-pytorch
Pillow
tqdm
torchvision
scikit-learn
streamlit